Create ReviewSafeTransactionViewController
