+++ Multisig/UI/Transaction/ExecuteTransaction/MiniAccountAndBalancePiece.swift
