+++ Multisig/UI/Transaction/ExecuteTransaction/MiniAccountAndBalancePiece.xib
